Fixed homework-3 issues (devdide Selection and Catch controllers)
